refactor(ptv3): move Detection3DModule to its own file
Relocate the Detection3DModule class from ptv3_trt.{hpp,cpp} into
detection3d_module.{hpp,cpp}, completing the split. With every module now
external, ptv3_trt.hpp drops the low-level kernel/postprocess/config includes it
only needed while the classes were inlined and keeps just the module headers and
its own dependencies.
